extract server and routes providers
