+++ core/src/main/java/io/knotx/server/KnotxServerVerticle.java
@@ -15,26 +15,15 @@
  */
 package io.knotx.server;
 
-import io.knotx.server.api.handler.RoutingHandlerFactory;
 import io.knotx.server.configuration.KnotxServerOptions;
-import io.knotx.server.configuration.RoutingOperationOptions;
-import io.netty.handler.codec.http.HttpResponseStatus;
-import io.reactivex.SingleSource;
 import io.vertx.core.Context;
 import io.vertx.core.Future;
 import io.vertx.core.Vertx;
 import io.vertx.core.logging.Logger;
 import io.vertx.core.logging.LoggerFactory;
-import io.vertx.reactivex.RxHelper;
 import io.vertx.reactivex.core.AbstractVerticle;
-import io.vertx.reactivex.core.http.HttpServer;
-import io.vertx.reactivex.core.http.HttpServerRequest;
 import io.vertx.reactivex.ext.web.Router;
 import io.vertx.reactivex.ext.web.api.contract.openapi3.OpenAPI3RouterFactory;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.ServiceLoader;
-import java.util.stream.Collectors;
 
 public class KnotxServerVerticle extends AbstractVerticle {
 
@@ -43,7 +32,6 @@ public class KnotxServerVerticle extends AbstractVerticle {
 
   private static final Logger LOGGER = LoggerFactory.getLogger(KnotxServerVerticle.class);
 
-  private static final HttpResponseStatus BAD_REQUEST = HttpResponseStatus.BAD_REQUEST;
 
   private KnotxServerOptions options;
 
@@ -58,18 +46,18 @@ public class KnotxServerVerticle extends AbstractVerticle {
     LOGGER.info("Starting <{}>", this.getClass().getSimpleName());
     LOGGER.info("Open API specification location [{}]",
         options.getRoutingSpecificationLocation());
-    validateRoutingOperations();
 
-    final SecurityProvider securityProvider = new SecurityProvider(vertx,
-        options.getSecurityHandlers());
+    RoutesProvider routesProvider = new RoutesProvider(vertx, options.getRoutingOperations());
+    SecurityProvider securityProvider = new SecurityProvider(vertx, options.getSecurityHandlers());
+    HttpServerProvider httpServerProvider = new HttpServerProvider(vertx, options);
 
     OpenAPI3RouterFactory.rxCreate(vertx, options.getRoutingSpecificationLocation())
         .doOnSuccess(securityProvider::configureSecurity)
-        .doOnSuccess(this::configureRouting)
+        .doOnSuccess(routesProvider::configureRouting)
         .doOnSuccess(OpenAPI3RouterFactory::mountServicesFromExtensions)
         .map(OpenAPI3RouterFactory::getRouter)
         .doOnSuccess(this::logRouterRoutes)
-        .flatMap(this::configureHttpServer)
+        .flatMap(httpServerProvider::configureHttpServer)
         .subscribe(
             ok -> {
               LOGGER.info("Knot.x HTTP Server started. Listening on port {}",
@@ -83,118 +71,6 @@ public class KnotxServerVerticle extends AbstractVerticle {
         );
   }
 
-  private void configureRouting(OpenAPI3RouterFactory routerFactory) {
-    List<RoutingHandlerFactory> handlerFactories = loadRoutingHandlerFactories();
-    options.getRoutingOperations().forEach(operation -> {
-      registerRoutingHandlersPerOperation(routerFactory, handlerFactories, operation);
-      registerFailureHandlersPerOperation(routerFactory, handlerFactories, operation);
-      LOGGER.info("Initialized all handlers for operation [{}]", operation.getOperationId());
-      if (LOGGER.isDebugEnabled()) {
-        LOGGER.debug("Operation initialization details [{}]", operation.toJson().encodePrettily());
-      }
-    });
-  }
-
-  private void validateRoutingOperations() {
-    if (options.getRoutingOperations() == null || options.getRoutingOperations().isEmpty()) {
-      LOGGER.warn(
-          "The server configuration does not contain any operation defined. Please check your "
-              + "configuration [config.server.options.config.routingOperations]");
-    }
-  }
-
-  private void registerRoutingHandlersPerOperation(OpenAPI3RouterFactory routerFactory,
-      List<RoutingHandlerFactory> handlerFactories,
-      RoutingOperationOptions operation) {
-    operation.getHandlers().forEach(routingHandlerOptions ->
-        handlerFactories.stream()
-            .filter(
-                handlerFactory -> handlerFactory.getName()
-                    .equals(routingHandlerOptions.getName()))
-            .findAny()
-            .map(handlerFactory ->
-                routerFactory
-                    .addHandlerByOperationId(operation.getOperationId(),
-                        handlerFactory.create(vertx, routingHandlerOptions.getConfig()))
-            )
-            .orElseThrow(() -> {
-              LOGGER.error(
-                  "Handler factory [{}] not found in registered factories [{}], all options [{}]",
-                  routingHandlerOptions.getName(), handlerFactories, operation);
-              return new IllegalStateException(
-                  "Can not find handler factory for [" + routingHandlerOptions.getName() + "]");
-            })
-    );
-  }
-
-  private void registerFailureHandlersPerOperation(OpenAPI3RouterFactory openAPI3RouterFactory,
-      List<RoutingHandlerFactory> handlerFactories,
-      RoutingOperationOptions routingOperationOptions) {
-    routingOperationOptions.getFailureHandlers().forEach(routingHandlerOptions ->
-        handlerFactories.stream()
-            .filter(
-                handlerFactory -> handlerFactory.getName()
-                    .equals(routingHandlerOptions.getName()))
-            .findAny()
-            .map(handlerFactory ->
-                openAPI3RouterFactory
-                    .addFailureHandlerByOperationId(routingOperationOptions.getOperationId(),
-                        handlerFactory.create(vertx, routingHandlerOptions.getConfig()))
-            )
-            .orElseThrow(IllegalStateException::new)
-    );
-  }
-
-  private SingleSource<? extends HttpServer> configureHttpServer(Router router) {
-    HttpServer httpServer = vertx.createHttpServer(options.getServerOptions());
-
-    if (options.isDropRequests()) {
-      httpServer.requestStream().toFlowable()
-          .map(HttpServerRequest::pause)
-          .onBackpressureBuffer(options.getBackpressureBufferCapacity(),
-              () -> LOGGER.warn("Backpressure buffer is overflown. Dropping request"),
-              options.getBackpressureStrategy())
-          .onBackpressureDrop(
-              req -> req.response().setStatusCode(options.getDropRequestResponseCode()).end())
-          .observeOn(RxHelper.scheduler(vertx.getDelegate()))
-          .subscribe(req -> {
-            req.resume();
-            routeSafe(req, router);
-          }, error -> LOGGER.error("Exception while processing!", error));
-    } else {
-      httpServer
-          .requestHandler(req -> routeSafe(req, router));
-    }
-
-    return httpServer.rxListen();
-  }
-
-  private void routeSafe(HttpServerRequest req, Router router) {
-    try {
-      router.accept(req);
-    } catch (IllegalArgumentException ex) {
-      LOGGER.warn("Problem decoding Query String ", ex);
-
-      req.response()
-          .setStatusCode(BAD_REQUEST.code())
-          .setStatusMessage(BAD_REQUEST.reasonPhrase())
-          .end("Invalid characters in Query Parameter");
-    }
-  }
-
-  private List<RoutingHandlerFactory> loadRoutingHandlerFactories() {
-    List<RoutingHandlerFactory> routingFactories = new ArrayList<>();
-    ServiceLoader.load(RoutingHandlerFactory.class)
-        .iterator()
-        .forEachRemaining(routingFactories::add);
-
-    LOGGER.info("Routing handler factory names [{}] registered.",
-        routingFactories.stream().map(RoutingHandlerFactory::getName).collect(Collectors
-            .joining(",")));
-
-    return routingFactories;
-  }
-
   private void logRouterRoutes(Router router) {
     LOGGER.info("Routes [{}]", router.getRoutes());
     printRoutes(router);
@@ -202,15 +78,11 @@ public class KnotxServerVerticle extends AbstractVerticle {
 
   private void printRoutes(Router router) {
     // @formatter:off
-    System.out.println(
-        "\n@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@");
-    System.out.println(
-        "@@                              ROUTER CONFIG                                 @@");
-    System.out.println(
-        "@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@");
+    System.out.println("\n@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@");
+    System.out.println("@@                              ROUTER CONFIG                                 @@");
+    System.out.println("@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@");
     router.getRoutes().forEach(route -> System.out.println("@@     " + route.getDelegate()));
-    System.out.println(
-        "@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n");
+    System.out.println("@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@\n");
     // @formatter:on
   }
